Network Update VERY IMPORTANT README
I have modified the client so it can connect to a server over the
network (in theory). There is a static boolean in GameClient that is
called LOCAL, right now it is set to false (indicating that the game is
run over the internet. The client will have to manually enter the IP
address of the host in the IP field when starting the game. When LOCAL
is true the client should not enter the IP address as this will be
defaulted to "localhost".